+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalLineItemKind.kt
@@ -1,0 +1,12 @@
+package com.braintreepayments.api.paypal
+
+/**
+ * The type of PayPal line item.
+ *
+ * [CREDIT] A line item that is a credit.
+ * [DEBIT] A line item that debits.
+ */
+enum class PayPalLineItemKind(val stringValue: String) {
+    CREDIT("credit"),
+    DEBIT("debit")
+}