+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalInternalClientUnitTest.java
@@ -6,6 +6,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
@@ -93,7 +94,7 @@ public class PayPalInternalClientUnitTest {
         PayPalVaultRequest payPalRequest = new PayPalVaultRequest(true);
         payPalRequest.setBillingAgreementDescription("Billing Agreement Description");
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setLandingPageType("sample-landing-page-type");
+        payPalRequest.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_BILLING);
         payPalRequest.setDisplayName("sample-display-name");
         payPalRequest.setLocaleCode("US");
         payPalRequest.setShippingAddressRequired(true);
@@ -104,8 +105,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(eq("/v1/paypal_hermes/setup_billing_agreement"),
-                captor.capture(), any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            eq("/v1/paypal_hermes/setup_billing_agreement"),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -118,7 +123,7 @@ public class PayPalInternalClientUnitTest {
                 .put("description", "Billing Agreement Description")
                 .put("experience_profile", new JSONObject()
                         .put("no_shipping", false)
-                        .put("landing_page_type", "sample-landing-page-type")
+                        .put("landing_page_type", "billing")
                         .put("brand_name", "sample-display-name")
                         .put("locale_code", "US")
                         .put("address_override", false))
@@ -155,7 +160,11 @@ public class PayPalInternalClientUnitTest {
         shippingAddressOverride.setPostalCode("12345");
         shippingAddressOverride.setCountryCodeAlpha2("US");
 
-        PayPalLineItem item = new PayPalLineItem(PayPalLineItem.KIND_DEBIT, "Item 0", "1", "2");
+        PayPalLineItem item = new PayPalLineItem(PayPalLineItemKind.DEBIT,
+            "Item 0",
+            "1",
+            "2"
+        );
         item.setDescription("A new item");
         item.setProductCode("abc-123");
         item.setUnitTaxAmount("1.50");
@@ -163,11 +172,11 @@ public class PayPalInternalClientUnitTest {
 
         PayPalCheckoutRequest payPalRequest = new PayPalCheckoutRequest("1.00", true);
         payPalRequest.setCurrencyCode("USD");
-        payPalRequest.setIntent("authorize");
+        payPalRequest.setIntent(PayPalPaymentIntent.AUTHORIZE);
         payPalRequest.setShouldRequestBillingAgreement(true);
         payPalRequest.setBillingAgreementDescription("Billing Agreement Description");
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setLandingPageType("sample-landing-page-type");
+        payPalRequest.setLandingPageType(PayPalLandingPageType.LANDING_PAGE_TYPE_LOGIN);
         payPalRequest.setDisplayName("sample-display-name");
         payPalRequest.setLocaleCode("US");
         payPalRequest.setShippingAddressRequired(true);
@@ -179,8 +188,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(eq("/v1/paypal_hermes/create_payment_resource"),
-                captor.capture(), any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            eq("/v1/paypal_hermes/create_payment_resource"),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -208,7 +221,7 @@ public class PayPalInternalClientUnitTest {
                                 .put("url", "http://example.com")))
                 .put("experience_profile", new JSONObject()
                         .put("no_shipping", false)
-                        .put("landing_page_type", "sample-landing-page-type")
+                        .put("landing_page_type", "login")
                         .put("brand_name", "sample-display-name")
                         .put("locale_code", "US")
                         .put("address_override", false))
@@ -238,8 +251,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -263,8 +280,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -287,8 +308,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -311,8 +336,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -335,8 +364,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -363,8 +396,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(eq("/v1/paypal_hermes/setup_billing_agreement"),
-                captor.capture(), any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            eq("/v1/paypal_hermes/setup_billing_agreement"),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -387,8 +424,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -410,8 +451,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -433,8 +478,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -508,8 +557,12 @@ public class PayPalInternalClientUnitTest {
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
 
         ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
-        verify(braintreeClient).sendPOST(anyString(), captor.capture(),
-                any(HttpResponseCallback.class));
+        verify(braintreeClient).sendPOST(
+            anyString(),
+            captor.capture(),
+            anyMap(),
+            any(HttpResponseCallback.class)
+        );
 
         String result = captor.getValue();
         JSONObject actual = new JSONObject(result);
@@ -626,9 +679,9 @@ public class PayPalInternalClientUnitTest {
         PayPalInternalClient sut = new PayPalInternalClient(braintreeClient, dataCollector, apiClient);
 
         PayPalCheckoutRequest payPalRequest = new PayPalCheckoutRequest("1.00", true);
-        payPalRequest.setIntent("authorize");
+        payPalRequest.setIntent(PayPalPaymentIntent.AUTHORIZE);
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
-        payPalRequest.setUserAction(PayPalCheckoutRequest.USER_ACTION_COMMIT);
+        payPalRequest.setUserAction(PayPalPaymentUserAction.USER_ACTION_COMMIT);
         payPalRequest.setRiskCorrelationId("sample-client-metadata-id");
 
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);
@@ -641,7 +694,7 @@ public class PayPalInternalClientUnitTest {
                 "https://checkout.paypal.com/one-touch-login-sandbox/index.html?token=fake-token&action=create_payment_resource&amount=1.00&authorization_fingerprint=63cc461306c35080ce674a3372bffe1580b4130c7fd33d33968aa76bb93cdd06%7Ccreated_at%3D2015-10-13T18%3A49%3A48.371382792%2B0000%26merchant_id%3Ddcpspy2brwdjr3qn%26public_key%3D9wwrzqk3vr3t4nc8&cancel_url=com.braintreepayments.api.test.braintree%3A%2F%2Fonetouch%2Fv1%2Fcancel&controller=client_api%2Fpaypal_hermes&currency_iso_code=USD&experience_profile%5Baddress_override%5D=false&experience_profile%5Bno_shipping%5D=false&merchant_id=dcpspy2brwdjr3qn&return_url=com.braintreepayments.api.test.braintree%3A%2F%2Fonetouch%2Fv1%2Fsuccess&offer_paypal_credit=true&version=1";
         PayPalPaymentAuthRequestParams payPalPaymentAuthRequestParams = captor.getValue();
         assertFalse(payPalPaymentAuthRequestParams.isBillingAgreement());
-        assertEquals("authorize", payPalPaymentAuthRequestParams.getIntent());
+        assertEquals(PayPalPaymentIntent.AUTHORIZE, payPalPaymentAuthRequestParams.getIntent());
         assertEquals("sample-merchant-account-id", payPalPaymentAuthRequestParams.getMerchantAccountId());
         assertEquals("https://example.com://onetouch/v1/success", payPalPaymentAuthRequestParams.getSuccessUrl());
         assertEquals("fake-token", payPalPaymentAuthRequestParams.getPairingId());
@@ -766,7 +819,7 @@ public class PayPalInternalClientUnitTest {
         PayPalInternalClient sut = new PayPalInternalClient(braintreeClient, dataCollector, apiClient);
 
         PayPalCheckoutRequest payPalRequest = new PayPalCheckoutRequest("1.00", true);
-        payPalRequest.setIntent("authorize");
+        payPalRequest.setIntent(PayPalPaymentIntent.AUTHORIZE);
         payPalRequest.setMerchantAccountId("sample-merchant-account-id");
 
         sut.sendRequest(context, payPalRequest, payPalInternalClientCallback);